+++ hanvest/hanvest/SharedUI/Button/Default/Enums/HanvestButtonDefaultSize.swift
@@ -1,0 +1,11 @@
+//
+//  HanvestButtonDefaultSize.swift
+//  hanvest
+//
+//  Created by Hans Arthur Cupiterson on 09/10/24.
+//
+
+enum HanvestButtonDefaultSize {
+    case large
+    case medium
+}
