+++ hanvest/hanvest/SharedUI/Button/Default/Enums/HanvestButtonDefaultState.swift
@@ -1,0 +1,12 @@
+//
+//  HanvestButtonDefaultState.swift
+//  hanvest
+//
+//  Created by Hans Arthur Cupiterson on 09/10/24.
+//
+
+enum HanvestButtonDefaultState {
+    case unpressed
+    case pressed
+    case disabled
+}
