+++ hanvest/hanvest/SharedUI/Button/Default/Enums/HanvestButtonDefaultStyle.swift
@@ -1,0 +1,12 @@
+//
+//  HanvestButtonDefaultStyle.swift
+//  hanvest
+//
+//  Created by Hans Arthur Cupiterson on 09/10/24.
+//
+
+enum HanvestButtonDefaultStyle {
+    case filled
+    case bordered
+    case borderless
+}
