--- hanvest/hanvest/SharedUI/Button/HanvestButtonDefault.swift
@@ -1,24 +1,0 @@
-//
-//  HanvestButtonDefault.swift
-//  hanvest
-//
-//  Created by Hans Arthur Cupiterson on 09/10/24.
-//
-
-import SwiftUI
-
-struct HanvestButtonDefault: View {
-    var 
-    
-    var title: String
-    var image: Image?
-    var action: () -> Void
-    
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-#Preview {
-    HanvestButtonDefault()
-}
